scraper/requirements.txt
Flask
requests
python-dotenv
playwright
beautifulsoup4
pyvirtualdisplay
fake-useragent
